+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 
 - Handle ``:identifier`` modifier for reference search parameters [simonvadee]
 
+- fixes `BundleWrapper`` as_json mode, now includes with ``resourceType`` value. [nazrulworld]
+
 - ``Dict`` response option has bee added in ``fhirpath.search.fhir_search`` [nazrulworld]
 
 - Ignore empty search params #21 [simonvadee]